Change the extension to work with WikiMedia 1.35 and greater
